Added the C167 processor interface.
This includes Alejandro Saez's original code that's been modified
slightly to remove the inout port and replace it with _I and _O
ports instead. The top level MHS contains the only inout as per
the Xilinx preference.

Also lumped in a new test point output that, TP_p, only 2 of which
are routed externally on the PIC.